requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
colorama>=0.4.4
Pillow>=8.0.0
moviepy>=1.0.3